JSON Values bigger than max integer value leaving columns out of output file #278

+++ src/ChoETL.JSON/ChoJSONRecordReader.cs
@@ -2185,7 +2185,7 @@ namespace ChoETL
                     case JTokenType.String:
                         return (string)jToken;
                     case JTokenType.Integer:
-                        return (int)jToken;
+                        return (long)jToken;
                     case JTokenType.Float:
                         return (float)jToken;
                     case JTokenType.Date:
